added a visualization script

+++ src/GEL/Geometry/rply.c
@@ -1224,13 +1224,13 @@ static int oascii_uint16(p_ply ply, double value) {
 }
 
 static int oascii_int32(p_ply ply, double value) {
-    if (value > LONG_MAX || value < LONG_MIN) return 0;
+    if (value > INT_MAX || value < INT_MIN) return 0;
     return fprintf(ply->fp, "%d ", (int) value) > 0;
 }
 
 static int oascii_uint32(p_ply ply, double value) {
-    if (value > ULONG_MAX || value < 0) return 0;
-    return fprintf(ply->fp, "%d ", (unsigned int) value) > 0;
+    if (value > UINT_MAX || value < 0) return 0;
+    return fprintf(ply->fp, "%u ", (unsigned int) value) > 0;
 }
 
 static int oascii_float32(p_ply ply, double value) {
@@ -1269,13 +1269,13 @@ static int obinary_uint16(p_ply ply, double value) {
 
 static int obinary_int32(p_ply ply, double value) {
     int int32 = (int) value;
-    if (value > LONG_MAX || value < LONG_MIN) return 0;
+    if (value > INT_MAX || value < INT_MIN) return 0;
     return ply->odriver->ochunk(ply, &int32, sizeof(int32));
 }
 
 static int obinary_uint32(p_ply ply, double value) {
     unsigned int uint32 = (unsigned int) value;
-    if (value > ULONG_MAX || value < 0) return 0;
+    if (value > UINT_MAX || value < 0) return 0;
     return ply->odriver->ochunk(ply, &uint32, sizeof(uint32));
 }
 
@@ -1328,7 +1328,7 @@ static int iascii_int32(p_ply ply, double *value) {
     char *end;
     if (!ply_read_word(ply)) return 0;
     *value = strtol(BWORD(ply), &end, 10);
-    if (*end || *value > LONG_MAX || *value < LONG_MIN) return 0;
+    if (*end || *value > INT_MAX || *value < INT_MIN) return 0;
     return 1;
 }
